+++ src/Twig/TemporaryUrlTwigExtension.php
@@ -32,7 +32,7 @@ class TemporaryUrlTwigExtension extends AbstractExtension
         ];
     }
 
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new TwigFunction(